# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BARQ Agentic Incident Resolution Platform",
    "description": "Autonomous incident triage, diagnosis, and remediation platform with HITL governance.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/approvals": {
      "get": {
        "tags": [
          "Approvals"
        ],
        "summary": "List all approval decisions",
        "description": "Fetch recorded human approval decisions, optionally filtered by execution ID or deciding operator.",
        "operationId": "list_approvals_api_v1_approvals_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "execution_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional filter by execution ID",
              "title": "Execution Id"
            },
            "description": "Optional filter by execution ID"
          },
          {
            "name": "decided_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional filter by deciding operator user ID",
              "title": "Decided By"
            },
            "description": "Optional filter by deciding operator user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApprovalResponse"
                  },
                  "title": "Response List Approvals Api V1 Approvals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals/{id}": {
      "get": {
        "tags": [
          "Approvals"
        ],
        "summary": "Get approval details by ID",
        "description": "Fetch a specific human approval record by its unique ID.",
        "operationId": "get_approval_api_v1_approvals__id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals/{id}/decide": {
      "post": {
        "tags": [
          "Approvals"
        ],
        "summary": "Submit human operator approval decision",
        "description": "Record or submit a human operator decision ('approved', 'rejected', 'cancelled', 'expired') for an approval request.",
        "operationId": "decide_approval_api_v1_approvals__id__decide_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dlq": {
      "get": {
        "tags": [
          "Dead-Letter Queue"
        ],
        "summary": "List all dead-lettered events",
        "description": "Retrieve all poisoned or retry-exhausted incident events currently in the DLQ.",
        "operationId": "list_dlq_events_api_v1_dlq_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DLQEventResponse"
                  },
                  "type": "array",
                  "title": "Response List Dlq Events Api V1 Dlq Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/dlq/{event_id}/replay": {
      "post": {
        "tags": [
          "Dead-Letter Queue"
        ],
        "summary": "Replay a dead-lettered event",
        "description": "Replay a poisoned or failed event from the DLQ. Enforces Operator Role RBAC.",
        "operationId": "replay_dlq_event_api_v1_dlq__event_id__replay_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique event ID of the dead-lettered event to replay",
              "title": "Event Id"
            },
            "description": "The unique event ID of the dead-lettered event to replay"
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DLQReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "rejected",
              "cancelled",
              "expired"
            ],
            "title": "Decision",
            "description": "Decision: 'approved', 'rejected', 'cancelled', or 'expired'"
          },
          "decided_by": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Decided By",
            "description": "Identity of the deciding operator"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Optional justification, feedback, or remediation note"
          },
          "evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence",
            "description": "Optional structured evidence or parameter overrides"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "decided_by"
        ],
        "title": "ApprovalDecisionRequest",
        "description": "Schema for submitting a human operator approval decision."
      },
      "ApprovalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Primary key from approvals.id"
          },
          "execution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Execution Id",
            "description": "Foreign key to executions.execution_id"
          },
          "workflow_state_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow State Id",
            "description": "Foreign key to workflow_state.id if associated with a specific node"
          },
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "rejected",
              "cancelled",
              "expired"
            ],
            "title": "Decision",
            "description": "Decision outcome matching approvals.decision constraint"
          },
          "decided_by": {
            "type": "string",
            "maxLength": 255,
            "title": "Decided By",
            "description": "Identity of the operator or automated service that decided"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Operator justification or explanation text"
          },
          "evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence",
            "description": "Structured context or parameters evaluated during decision"
          },
          "decided_at": {
            "type": "string",
            "format": "date-time",
            "title": "Decided At",
            "description": "Timestamp when the decision was finalized"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "execution_id",
          "decision",
          "decided_by",
          "decided_at"
        ],
        "title": "ApprovalResponse",
        "description": "Schema representing an approval record from the 'approvals' table."
      },
      "DLQEventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id",
            "description": "Unique event identifier"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload",
            "description": "Raw outbound event payload that failed processing"
          },
          "failure_reason": {
            "type": "string",
            "title": "Failure Reason",
            "description": "Reason or error message that routed the event to DLQ"
          },
          "retry_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Retry Count",
            "description": "Number of retry attempts attempted before landing in DLQ",
            "default": 0
          },
          "failed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Failed At",
            "description": "Timestamp when event was placed into the DLQ"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "payload",
          "failure_reason",
          "failed_at"
        ],
        "title": "DLQEventResponse",
        "description": "Schema representing a poisoned or retry-exhausted incident event in the DLQ."
      },
      "DLQReplayResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id",
            "description": "Identifier of the event being replayed"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "replayed"
            ],
            "title": "Status",
            "description": "Replay outcome status ('accepted' or 'replayed')",
            "default": "accepted"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id",
            "description": "Correlation ID tracking the replay operation"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable result summary of the replay dispatch"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "correlation_id",
          "message"
        ],
        "title": "DLQReplayResponse",
        "description": "Response schema returned upon replaying a DLQ event."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Bearer token authentication. Enter your token (without 'Bearer ' prefix).",
        "scheme": "bearer"
      }
    }
  }
}